requirements.txt
numpy>=2.2.0
matplotlib>=3.10.0
tifffile>=2025.2.0
Pillow>=11.1.0
tqdm>=4.66.0
